dissector: split out RPM extraction to avoid too many open files
When we open the SRPM files in the dissector in order to extract their
contents, we use defer to ensure they get closed later, but that doesn't
happen until the function returns, which means that there could be
hundreds of files open at a time. On my system with all packages that
results in the program terminating with "too many open files". If we
move the extraction of a single package to its own function, then the
deferred close happens at the end of processing each package instead.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>