dissector: re-use previously downloaded SRPMs
Clear Linux releases quite often, but many of the SRPMs are identical
between releases where no changes were required, thus a mechanism to
re-use previously downloaded packages that haven't changed would save
quite a bit of time. Once we've successfully downloaded the SRPMs, create
a .done file in the srpms subdirectory indicating that we have finished.
The next time we run dissector, we look for the most recent subdirectory
with the .done file with a version number less than the requested
version, and copy in any SRPMs from there to the new version directory
where names and checksums match the new ones, using hard linking to
save space. (When we subsequently get to the download stage, the files
that already exist will be skipped as per existing logic.)

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>